+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroPluginExtension.kt
@@ -55,9 +55,6 @@ constructor(
   public val generateAssistedFactories: Property<Boolean> =
     objects.booleanProperty("metro.generateAssistedFactories", false)
 
-  @Deprecated("This does not do anything anymore and will be removed")
-  public abstract val generateThrowsAnnotations: Property<Boolean>
-
   /**
    * Enables whether the Metro compiler plugin can inject top-level functions. See the kdoc on
    * `Inject` for more details.
@@ -180,11 +177,6 @@ constructor(
   public val shrinkUnusedBindings: Property<Boolean> =
     objects.booleanProperty("metro.shrinkUnusedBindings", true)
 
-  /** Enable/disable chunking of field initializers. Enabled by default. */
-  @Deprecated("This is the default and the option will be removed", level = DeprecationLevel.ERROR)
-  public val chunkFieldInits: Property<Boolean> =
-    objects.booleanProperty("metro.chunkFieldInits", true)
-
   /**
    * Maximum number of statements per init function when chunking field initializers. Default is 25,
    * must be > 0.
@@ -224,21 +216,6 @@ constructor(
     objects
       .property(OptionalBindingBehavior::class.java)
       .convention(OptionalBindingBehavior.DEFAULT)
-
-  /** Enable/disable automatic transformation of providers to be private. Enabled by default. */
-  @Deprecated(
-    "Transforming providers to private is deprecated as it results in less efficient code generation",
-    level = DeprecationLevel.ERROR,
-  )
-  public val transformProvidersToPrivate: Property<Boolean> =
-    objects.booleanProperty("metro.transformProvidersToPrivate", false)
-
-  @Deprecated(
-    "This is just a proxy to publicScopedProviderSeverity now",
-    replaceWith = ReplaceWith("publicScopedProviderSeverity"),
-  )
-  public val publicProviderSeverity: Property<DiagnosticSeverity>
-    get() = publicScopedProviderSeverity
 
   /**
    * Configures the Metro compiler plugin to warn, error, or do nothing when it encounters
@@ -390,12 +367,6 @@ constructor(
    */
   public val useAssistedParamNamesAsIdentifiers: Property<Boolean> =
     objects.booleanProperty("metro.useAssistedParamNamesAsIdentifiers", true)
-
-  @Deprecated(
-    "The `Assisted.value` property is now formally deprecated and this control no longer does anything."
-  )
-  public val assistedIdentifierSeverity: Property<DiagnosticSeverity> =
-    objects.enumProperty<DiagnosticSeverity>("assistedIdentifierSeverity", DiagnosticSeverity.WARN)
 
   /**
    * Compiler version aliases mapping fake IDE versions to their real compiler versions.